added swap logic

+++ contracts/factory/src/state.rs
@@ -1,16 +1,15 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use cosmwasm_std::{CanonicalAddr, Addr, Coin, Uint128};
+use cosmwasm_std::{CanonicalAddr, Addr, Uint128};
 use cw_storage_plus::{Item, Map};
-use cw20::{Balance, Cw20CoinVerified};
 
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct State {
     pub admin: CanonicalAddr,
     pub otc_code_hash: u64,
-    pub index: u64,
+    pub index: u32,
     pub active: bool,
 }
